Refactored library and added modules for processing grid in json format
