+++ src/components/GridLayout/PageLoaderBackup.vue
@@ -1,0 +1,35 @@
+<template>
+<div class="columns is-vcentered is-centered" style="height: 100vh">
+  <div class="column is-half">
+      <div class="loader"></div>
+
+  </div>
+</div>
+</template>
+
+<style scoped>
+
+.loader {
+  border: 16px solid #f3f3f3;
+  border-radius: 50%;
+  border-top: 16px solid #3498db;
+  width: 120px;
+  height: 120px;
+  -webkit-animation: spin 2s linear infinite; /* Safari */
+  animation: spin 2s linear infinite;
+}
+
+/* Safari */
+@-webkit-keyframes spin {
+  0% { -webkit-transform: rotate(0deg); }
+  100% { -webkit-transform: rotate(360deg); }
+}
+
+@keyframes spin {
+  0% { transform: rotate(0deg); }
+  100% { transform: rotate(360deg); }
+}
+
+
+
+</style>
