+++ src/lib/GridParser/CellClassStyle/index.js
@@ -1,0 +1,19 @@
+/*
+Appends function to class rules if empty for each row
+
+editable_pass_style
+editable_error_style
+non_editable_pass_style
+non_editable_error_style
+
+*/
+
+class CellClassInit {
+    constructor(json_grid) {
+        this.jg = json_grid
+    }
+    Run() {
+        //assembles grid
+    }
+
+}
