+++ src/lib/GridParser/ExpressionParser/ToolTips/index.js
@@ -1,0 +1,1 @@
+//creates tool tip from expression syntax
