+++ src/lib/GridParser/ExpressionParser/index.js
@@ -1,0 +1,30 @@
+/*
+Handles function processing
+
+
+IsError
+ValueGetter
+
+params
+
+*/
+
+//AddFunctions i.e. concat
+//AddGlobals i.e. defaults
+//Concatenate
+//IfNull
+//IfNullOrUndefined
+
+import math from 'mathjs'
+//bignumber
+//regular
+
+class ExpressionParser {
+    constructor(json_grid, global_params) {
+        this.jg = json_grid
+        this.gp = global_params
+    }
+    CreateErrorFunction() {}
+    CreateValueGetterFunction(){}
+    CreateFunction() {}
+}
