+++ src/lib/InitialParams/README
@@ -1,0 +1,3 @@
+pulls data
+sets initial sort, order_by values
+creates initial global object for calculations
